server/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
dnspython==2.4.2
executing==2.0.0
faiss-cpu==1.7.4
fastapi==0.104.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-vision==3.4.5
googleapis-common-protos==1.61.0
gradio_client==0.7.1
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.17.3
ibm-cos-sdk==2.13.2
ibm-cos-sdk-core==2.13.2
ibm-cos-sdk-s3transfer==2.13.2
ibm-watson-machine-learning==1.0.327
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.4.0
jupyter_core==5.4.0
loguru==0.7.2
lomond==0.3.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.2
numpy==1.26.1
packaging==23.2
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==3.11.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
safetensors==0.4.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.1
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.18.0
vision-transformer-pytorch==1.0.3
watchfiles==0.21.0
wcwidth==0.2.8
websockets==11.0.3
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
